library: add module_utils namespace
This adds a new `module_utils` namespace in order to avoid defining same
functions in each module.
