+++ ansible.cfg
@@ -4,6 +4,7 @@
 [defaults]
 ansible_managed = Please do not change this file directly since it is managed by Ansible and will be overwritten
 library = ./library
+module_utils = ./module_utils
 action_plugins = plugins/actions
 callback_plugins = plugins/callback
 filter_plugins = plugins/filter
