+++ library/ceph_fs.py
@@ -16,9 +16,19 @@ from __future__ import absolute_import, division, print_function
 __metaclass__ = type
 
 from ansible.module_utils.basic import AnsibleModule
+try:
+    from ansible.module_utils.ca_common import is_containerized, \
+                                               exec_command, \
+                                               generate_ceph_cmd, \
+                                               exit_module
+except ImportError:
+    from module_utils.ca_common import is_containerized, \
+                                       exec_command, \
+                                       generate_ceph_cmd, \
+                                       exit_module
+
 import datetime
 import json
-import os
 
 
 ANSIBLE_METADATA = {
@@ -97,76 +107,6 @@ EXAMPLES = '''
 RETURN = '''#  '''
 
 
-def container_exec(binary, container_image):
-    '''
-    Build the docker CLI to run a command inside a container
-    '''
-
-    container_binary = os.getenv('CEPH_CONTAINER_BINARY')
-    command_exec = [container_binary,
-                    'run',
-                    '--rm',
-                    '--net=host',
-                    '-v', '/etc/ceph:/etc/ceph:z',
-                    '-v', '/var/lib/ceph/:/var/lib/ceph/:z',
-                    '-v', '/var/log/ceph/:/var/log/ceph/:z',
-                    '--entrypoint=' + binary, container_image]
-    return command_exec
-
-
-def is_containerized():
-    '''
-    Check if we are running on a containerized cluster
-    '''
-
-    if 'CEPH_CONTAINER_IMAGE' in os.environ:
-        container_image = os.getenv('CEPH_CONTAINER_IMAGE')
-    else:
-        container_image = None
-
-    return container_image
-
-
-def pre_generate_ceph_cmd(container_image=None):
-    '''
-    Generate ceph prefix comaand
-    '''
-    if container_image:
-        cmd = container_exec('ceph', container_image)
-    else:
-        cmd = ['ceph']
-
-    return cmd
-
-
-def generate_ceph_cmd(cluster, args, container_image=None):
-    '''
-    Generate 'ceph' command line to execute
-    '''
-
-    cmd = pre_generate_ceph_cmd(container_image=container_image)
-
-    base_cmd = [
-        '--cluster',
-        cluster,
-        'fs'
-    ]
-
-    cmd.extend(base_cmd + args)
-
-    return cmd
-
-
-def exec_commands(module, cmd):
-    '''
-    Execute command(s)
-    '''
-
-    rc, out, err = module.run_command(cmd)
-
-    return rc, cmd, out, err
-
-
 def create_fs(module, container_image=None):
     '''
     Create a new fs
@@ -179,7 +119,7 @@ def create_fs(module, container_image=None):
 
     args = ['new', name, metadata, data]
 
-    cmd = generate_ceph_cmd(cluster=cluster, args=args, container_image=container_image)
+    cmd = generate_ceph_cmd(cluster=cluster, sub_cmd=['fs'], args=args, container_image=container_image)
 
     return cmd
 
@@ -194,7 +134,7 @@ def get_fs(module, container_image=None):
 
     args = ['get', name, '--format=json']
 
-    cmd = generate_ceph_cmd(cluster=cluster, args=args, container_image=container_image)
+    cmd = generate_ceph_cmd(cluster=cluster, sub_cmd=['fs'], args=args, container_image=container_image)
 
     return cmd
 
@@ -209,7 +149,7 @@ def remove_fs(module, container_image=None):
 
     args = ['rm', name, '--yes-i-really-mean-it']
 
-    cmd = generate_ceph_cmd(cluster=cluster, args=args, container_image=container_image)
+    cmd = generate_ceph_cmd(cluster=cluster, sub_cmd=['fs'], args=args, container_image=container_image)
 
     return cmd
 
@@ -224,7 +164,7 @@ def fail_fs(module, container_image=None):
 
     args = ['fail', name]
 
-    cmd = generate_ceph_cmd(cluster=cluster, args=args, container_image=container_image)
+    cmd = generate_ceph_cmd(cluster=cluster, sub_cmd=['fs'], args=args, container_image=container_image)
 
     return cmd
 
@@ -240,26 +180,9 @@ def set_fs(module, container_image=None):
 
     args = ['set', name, 'max_mds', str(max_mds)]
 
-    cmd = generate_ceph_cmd(cluster=cluster, args=args, container_image=container_image)
+    cmd = generate_ceph_cmd(cluster=cluster, sub_cmd=['fs'], args=args, container_image=container_image)
 
     return cmd
-
-
-def exit_module(module, out, rc, cmd, err, startd, changed=False):
-    endd = datetime.datetime.now()
-    delta = endd - startd
-
-    result = dict(
-        cmd=cmd,
-        start=str(startd),
-        end=str(endd),
-        delta=str(delta),
-        rc=rc,
-        stdout=out.rstrip("\r\n"),
-        stderr=err.rstrip("\r\n"),
-        changed=changed,
-    )
-    module.exit_json(**result)
 
 
 def run_module():
@@ -301,25 +224,25 @@ def run_module():
     container_image = is_containerized()
 
     if state == "present":
-        rc, cmd, out, err = exec_commands(module, get_fs(module, container_image=container_image))
+        rc, cmd, out, err = exec_command(module, get_fs(module, container_image=container_image))
         if rc == 0:
             fs = json.loads(out)
             if max_mds and fs["mdsmap"]["max_mds"] != max_mds:
-                rc, cmd, out, err = exec_commands(module, set_fs(module, container_image=container_image))
+                rc, cmd, out, err = exec_command(module, set_fs(module, container_image=container_image))
                 if rc == 0:
                     changed = True
         else:
-            rc, cmd, out, err = exec_commands(module, create_fs(module, container_image=container_image))
+            rc, cmd, out, err = exec_command(module, create_fs(module, container_image=container_image))
             if max_mds and max_mds > 1:
-                exec_commands(module, set_fs(module, container_image=container_image))
+                exec_command(module, set_fs(module, container_image=container_image))
             if rc == 0:
                 changed = True
 
     elif state == "absent":
-        rc, cmd, out, err = exec_commands(module, get_fs(module, container_image=container_image))
+        rc, cmd, out, err = exec_command(module, get_fs(module, container_image=container_image))
         if rc == 0:
-            exec_commands(module, fail_fs(module, container_image=container_image))
-            rc, cmd, out, err = exec_commands(module, remove_fs(module, container_image=container_image))
+            exec_command(module, fail_fs(module, container_image=container_image))
+            rc, cmd, out, err = exec_command(module, remove_fs(module, container_image=container_image))
             if rc == 0:
                 changed = True
         else:
@@ -327,7 +250,7 @@ def run_module():
             out = "Ceph File System {} doesn't exist".format(name)
 
     elif state == "info":
-        rc, cmd, out, err = exec_commands(module, get_fs(module, container_image=container_image))
+        rc, cmd, out, err = exec_command(module, get_fs(module, container_image=container_image))
 
     exit_module(module=module, out=out, rc=rc, cmd=cmd, err=err, startd=startd, changed=changed)
