+++ library/ceph_pool.py
@@ -18,6 +18,20 @@ from __future__ import absolute_import, division, print_function
 __metaclass__ = type
 
 from ansible.module_utils.basic import AnsibleModule
+try:
+    from ansible.module_utils.ca_common import generate_ceph_cmd, \
+                                               pre_generate_ceph_cmd, \
+                                               is_containerized, \
+                                               exec_command, \
+                                               exit_module
+except ImportError:
+    from module_utils.ca_common import generate_ceph_cmd, \
+                                       pre_generate_ceph_cmd, \
+                                       is_containerized, \
+                                       exec_command, \
+                                       exit_module
+
+
 import datetime
 import json
 import os
@@ -141,81 +155,6 @@ pools:
 RETURN = '''#  '''
 
 
-def container_exec(binary, container_image):
-    '''
-    Build the docker CLI to run a command inside a container
-    '''
-
-    container_binary = os.getenv('CEPH_CONTAINER_BINARY')
-    command_exec = [container_binary,
-                    'run',
-                    '--rm',
-                    '--net=host',
-                    '-v', '/etc/ceph:/etc/ceph:z',
-                    '-v', '/var/lib/ceph/:/var/lib/ceph/:z',
-                    '-v', '/var/log/ceph/:/var/log/ceph/:z',
-                    '--entrypoint=' + binary, container_image]
-    return command_exec
-
-
-def is_containerized():
-    '''
-    Check if we are running on a containerized cluster
-    '''
-
-    if 'CEPH_CONTAINER_IMAGE' in os.environ:
-        container_image = os.getenv('CEPH_CONTAINER_IMAGE')
-    else:
-        container_image = None
-
-    return container_image
-
-
-def pre_generate_ceph_cmd(container_image=None):
-    if container_image:
-        binary = 'ceph'
-        cmd = container_exec(
-            binary, container_image)
-    else:
-        binary = ['ceph']
-        cmd = binary
-
-    return cmd
-
-
-def generate_ceph_cmd(cluster, args, user, user_key, container_image=None):
-    '''
-    Generate 'ceph' command line to execute
-    '''
-
-    cmd = pre_generate_ceph_cmd(container_image=container_image)
-
-    base_cmd = [
-        '-n',
-        user,
-        '-k',
-        user_key,
-        '--cluster',
-        cluster,
-        'osd',
-        'pool'
-    ]
-
-    cmd.extend(base_cmd + args)
-
-    return cmd
-
-
-def exec_commands(module, cmd):
-    '''
-    Execute command(s)
-    '''
-
-    rc, out, err = module.run_command(cmd)
-
-    return rc, cmd, out, err
-
-
 def check_pool_exist(cluster,
                      name,
                      user,
@@ -229,6 +168,7 @@ def check_pool_exist(cluster,
     args = ['stats', name, '-f', output_format]
 
     cmd = generate_ceph_cmd(cluster=cluster,
+                            sub_cmd=['osd', 'pool'],
                             args=args,
                             user=user,
                             user_key=user_key,
@@ -272,6 +212,7 @@ def get_application_pool(cluster,
     args = ['application', 'get', name, '-f', output_format]
 
     cmd = generate_ceph_cmd(cluster=cluster,
+                            sub_cmd=['osd', 'pool'],
                             args=args,
                             user=user,
                             user_key=user_key,
@@ -293,6 +234,7 @@ def enable_application_pool(cluster,
     args = ['application', 'enable', name, application]
 
     cmd = generate_ceph_cmd(cluster=cluster,
+                            sub_cmd=['osd', 'pool'],
                             args=args,
                             user=user,
                             user_key=user_key,
@@ -315,6 +257,7 @@ def disable_application_pool(cluster,
             application, '--yes-i-really-mean-it']
 
     cmd = generate_ceph_cmd(cluster=cluster,
+                            sub_cmd=['osd', 'pool'],
                             args=args,
                             user=user,
                             user_key=user_key,
@@ -337,22 +280,23 @@ def get_pool_details(module,
     args = ['ls', 'detail', '-f', output_format]
 
     cmd = generate_ceph_cmd(cluster=cluster,
+                            sub_cmd=['osd', 'pool'],
                             args=args,
                             user=user,
                             user_key=user_key,
                             container_image=container_image)
 
-    rc, cmd, out, err = exec_commands(module, cmd)
+    rc, cmd, out, err = exec_command(module, cmd)
 
     if rc == 0:
         out = [p for p in json.loads(out.strip()) if p['pool_name'] == name][0]
 
-    _rc, _cmd, application_pool, _err = exec_commands(module,
-                                                      get_application_pool(cluster,    # noqa: E501
-                                                                           name,    # noqa: E501
-                                                                           user,    # noqa: E501
-                                                                           user_key,    # noqa: E501
-                                                                           container_image=container_image))  # noqa: E501
+    _rc, _cmd, application_pool, _err = exec_command(module,
+                                                     get_application_pool(cluster,    # noqa: E501
+                                                                          name,    # noqa: E501
+                                                                          user,    # noqa: E501
+                                                                          user_key,    # noqa: E501
+                                                                          container_image=container_image))  # noqa: E501
 
     # This is a trick because "target_size_ratio" isn't present at the same level in the dict
     # ie:
@@ -424,6 +368,7 @@ def list_pools(cluster,
     args.extend(['-f', output_format])
 
     cmd = generate_ceph_cmd(cluster=cluster,
+                            sub_cmd=['osd', 'pool'],
                             args=args,
                             user=user,
                             user_key=user_key,
@@ -474,6 +419,7 @@ def create_pool(cluster,
                      user_pool_config['pg_autoscale_mode']['value']])
 
     cmd = generate_ceph_cmd(cluster=cluster,
+                            sub_cmd=['osd', 'pool'],
                             args=args,
                             user=user,
                             user_key=user_key,
@@ -490,6 +436,7 @@ def remove_pool(cluster, name, user, user_key, container_image=None):
     args = ['rm', name, name, '--yes-i-really-really-mean-it']
 
     cmd = generate_ceph_cmd(cluster=cluster,
+                            sub_cmd=['osd', 'pool'],
                             args=args,
                             user=user,
                             user_key=user_key,
@@ -514,21 +461,22 @@ def update_pool(module, cluster, name,
                     delta[key]['value']]
 
             cmd = generate_ceph_cmd(cluster=cluster,
+                                    sub_cmd=['osd', 'pool'],
                                     args=args,
                                     user=user,
                                     user_key=user_key,
                                     container_image=container_image)
 
-            rc, cmd, out, err = exec_commands(module, cmd)
+            rc, cmd, out, err = exec_command(module, cmd)
             if rc != 0:
                 return rc, cmd, out, err
 
         else:
-            rc, cmd, out, err = exec_commands(module, disable_application_pool(cluster, name, delta['application']['old_application'], user, user_key, container_image=container_image))  # noqa: E501
+            rc, cmd, out, err = exec_command(module, disable_application_pool(cluster, name, delta['application']['old_application'], user, user_key, container_image=container_image))  # noqa: E501
             if rc != 0:
                 return rc, cmd, out, err
 
-            rc, cmd, out, err = exec_commands(module, enable_application_pool(cluster, name, delta['application']['new_application'], user, user_key, container_image=container_image))  # noqa: E501
+            rc, cmd, out, err = exec_command(module, enable_application_pool(cluster, name, delta['application']['new_application'], user, user_key, container_image=container_image))  # noqa: E501
             if rc != 0:
                 return rc, cmd, out, err
 
@@ -536,23 +484,6 @@ def update_pool(module, cluster, name,
 
     out = report
     return rc, cmd, out, err
-
-
-def exit_module(module, out, rc, cmd, err, startd, changed=False):
-    endd = datetime.datetime.now()
-    delta = endd - startd
-
-    result = dict(
-        cmd=cmd,
-        start=str(startd),
-        end=str(endd),
-        delta=str(delta),
-        rc=rc,
-        stdout=out.rstrip("\r\n"),
-        stderr=err.rstrip("\r\n"),
-        changed=changed,
-    )
-    module.exit_json(**result)
 
 
 def run_module():
@@ -656,12 +587,12 @@ def run_module():
     user_key = os.path.join("/etc/ceph/", keyring_filename)
 
     if state == "present":
-        rc, cmd, out, err = exec_commands(module,
-                                          check_pool_exist(cluster,
-                                                           name,
-                                                           user,
-                                                           user_key,
-                                                           container_image=container_image))  # noqa: E501
+        rc, cmd, out, err = exec_command(module,
+                                         check_pool_exist(cluster,
+                                                          name,
+                                                          user,
+                                                          user_key,
+                                                          container_image=container_image))  # noqa: E501
         if rc == 0:
             running_pool_details = get_pool_details(module,
                                                     cluster,
@@ -696,49 +627,49 @@ def run_module():
             else:
                 out = "Pool {} already exists and there is nothing to update.".format(name)  # noqa: E501
         else:
-            rc, cmd, out, err = exec_commands(module,
-                                              create_pool(cluster,
-                                                          name,
-                                                          user,
-                                                          user_key,
-                                                          user_pool_config=user_pool_config,  # noqa: E501
-                                                          container_image=container_image))  # noqa: E501
+            rc, cmd, out, err = exec_command(module,
+                                             create_pool(cluster,
+                                                         name,
+                                                         user,
+                                                         user_key,
+                                                         user_pool_config=user_pool_config,  # noqa: E501
+                                                         container_image=container_image))  # noqa: E501
             if user_pool_config['application']['value']:
-                rc, _, _, _ = exec_commands(module,
-                                            enable_application_pool(cluster,
-                                                                    name,
-                                                                    user_pool_config['application']['value'],  # noqa: E501
-                                                                    user,
-                                                                    user_key,
-                                                                    container_image=container_image))  # noqa: E501
+                rc, _, _, _ = exec_command(module,
+                                           enable_application_pool(cluster,
+                                                                   name,
+                                                                   user_pool_config['application']['value'],  # noqa: E501
+                                                                   user,
+                                                                   user_key,
+                                                                   container_image=container_image))  # noqa: E501
             if user_pool_config['min_size']['value']:
                 # not implemented yet
                 pass
             changed = True
 
     elif state == "list":
-        rc, cmd, out, err = exec_commands(module,
-                                          list_pools(cluster,
-                                                     name, user,
-                                                     user_key,
-                                                     details,
-                                                     container_image=container_image))  # noqa: E501
+        rc, cmd, out, err = exec_command(module,
+                                         list_pools(cluster,
+                                                    name, user,
+                                                    user_key,
+                                                    details,
+                                                    container_image=container_image))  # noqa: E501
         if rc != 0:
             out = "Couldn't list pool(s) present on the cluster"
 
     elif state == "absent":
-        rc, cmd, out, err = exec_commands(module,
-                                          check_pool_exist(cluster,
-                                                           name, user,
-                                                           user_key,
-                                                           container_image=container_image))  # noqa: E501
-        if rc == 0:
-            rc, cmd, out, err = exec_commands(module,
-                                              remove_pool(cluster,
-                                                          name,
-                                                          user,
+        rc, cmd, out, err = exec_command(module,
+                                         check_pool_exist(cluster,
+                                                          name, user,
                                                           user_key,
                                                           container_image=container_image))  # noqa: E501
+        if rc == 0:
+            rc, cmd, out, err = exec_command(module,
+                                             remove_pool(cluster,
+                                                         name,
+                                                         user,
+                                                         user_key,
+                                                         container_image=container_image))  # noqa: E501
             changed = True
         else:
             rc = 0
