+++ tests/library/test_ceph_dashboard_user.py
@@ -1,9 +1,5 @@
-import os
-import sys
-from mock.mock import patch, MagicMock
-import pytest
-sys.path.append('./library')
-import ceph_dashboard_user  # noqa: E402
+from mock.mock import MagicMock
+import ceph_dashboard_user
 
 fake_binary = 'ceph'
 fake_cluster = 'ceph'
@@ -31,48 +27,13 @@ fake_params = {'cluster': fake_cluster,
 
 class TestCephDashboardUserModule(object):
 
-    @patch.dict(os.environ, {'CEPH_CONTAINER_BINARY': fake_container_binary})
-    def test_container_exec(self):
-        cmd = ceph_dashboard_user.container_exec(fake_binary, fake_container_image)
-        assert cmd == fake_container_cmd
-
-    def test_not_is_containerized(self):
-        assert ceph_dashboard_user.is_containerized() is None
-
-    @patch.dict(os.environ, {'CEPH_CONTAINER_IMAGE': fake_container_image})
-    def test_is_containerized(self):
-        assert ceph_dashboard_user.is_containerized() == fake_container_image
-
-    @pytest.mark.parametrize('image', [None, fake_container_image])
-    @patch.dict(os.environ, {'CEPH_CONTAINER_BINARY': fake_container_binary})
-    def test_pre_generate_ceph_cmd(self, image):
-        if image:
-            expected_cmd = fake_container_cmd
-        else:
-            expected_cmd = [fake_binary]
-
-        assert ceph_dashboard_user.pre_generate_ceph_cmd(image) == expected_cmd
-
-    @pytest.mark.parametrize('image', [None, fake_container_image])
-    @patch.dict(os.environ, {'CEPH_CONTAINER_BINARY': fake_container_binary})
-    def test_generate_ceph_cmd(self, image):
-        if image:
-            expected_cmd = fake_container_cmd
-        else:
-            expected_cmd = [fake_binary]
-
-        expected_cmd.extend([
-            '--cluster',
-            fake_cluster,
-            'dashboard'
-        ])
-        assert ceph_dashboard_user.generate_ceph_cmd(fake_cluster, [], image) == expected_cmd
-
     def test_create_user(self):
         fake_module = MagicMock()
         fake_module.params = fake_params
         expected_cmd = [
             fake_binary,
+            '-n', 'client.admin',
+            '-k', '/etc/ceph/ceph.client.admin.keyring',
             '--cluster', fake_cluster,
             'dashboard', 'ac-user-create',
             fake_user,
@@ -86,6 +47,8 @@ class TestCephDashboardUserModule(object):
         fake_module.params = fake_params
         expected_cmd = [
             fake_binary,
+            '-n', 'client.admin',
+            '-k', '/etc/ceph/ceph.client.admin.keyring',
             '--cluster', fake_cluster,
             'dashboard', 'ac-user-set-roles',
             fake_user
@@ -99,6 +62,8 @@ class TestCephDashboardUserModule(object):
         fake_module.params = fake_params
         expected_cmd = [
             fake_binary,
+            '-n', 'client.admin',
+            '-k', '/etc/ceph/ceph.client.admin.keyring',
             '--cluster', fake_cluster,
             'dashboard', 'ac-user-set-password',
             fake_user,
@@ -112,6 +77,8 @@ class TestCephDashboardUserModule(object):
         fake_module.params = fake_params
         expected_cmd = [
             fake_binary,
+            '-n', 'client.admin',
+            '-k', '/etc/ceph/ceph.client.admin.keyring',
             '--cluster', fake_cluster,
             'dashboard', 'ac-user-show',
             fake_user,
@@ -125,6 +92,8 @@ class TestCephDashboardUserModule(object):
         fake_module.params = fake_params
         expected_cmd = [
             fake_binary,
+            '-n', 'client.admin',
+            '-k', '/etc/ceph/ceph.client.admin.keyring',
             '--cluster', fake_cluster,
             'dashboard', 'ac-user-delete',
             fake_user
