+++ tests/library/test_ceph_fs.py
@@ -1,9 +1,5 @@
-import os
-import sys
-from mock.mock import patch, MagicMock
-import pytest
-sys.path.append('./library')
-import ceph_fs  # noqa : E402
+from mock.mock import MagicMock
+import ceph_fs
 
 
 fake_binary = 'ceph'
@@ -34,48 +30,13 @@ fake_params = {'cluster': fake_cluster,
 
 class TestCephFsModule(object):
 
-    @patch.dict(os.environ, {'CEPH_CONTAINER_BINARY': fake_container_binary})
-    def test_container_exec(self):
-        cmd = ceph_fs.container_exec(fake_binary, fake_container_image)
-        assert cmd == fake_container_cmd
-
-    def test_not_is_containerized(self):
-        assert ceph_fs.is_containerized() is None
-
-    @patch.dict(os.environ, {'CEPH_CONTAINER_IMAGE': fake_container_image})
-    def test_is_containerized(self):
-        assert ceph_fs.is_containerized() == fake_container_image
-
-    @pytest.mark.parametrize('image', [None, fake_container_image])
-    @patch.dict(os.environ, {'CEPH_CONTAINER_BINARY': fake_container_binary})
-    def test_pre_generate_ceph_cmd(self, image):
-        if image:
-            expected_cmd = fake_container_cmd
-        else:
-            expected_cmd = [fake_binary]
-
-        assert ceph_fs.pre_generate_ceph_cmd(image) == expected_cmd
-
-    @pytest.mark.parametrize('image', [None, fake_container_image])
-    @patch.dict(os.environ, {'CEPH_CONTAINER_BINARY': fake_container_binary})
-    def test_generate_ceph_cmd(self, image):
-        if image:
-            expected_cmd = fake_container_cmd
-        else:
-            expected_cmd = [fake_binary]
-
-        expected_cmd.extend([
-            '--cluster',
-            fake_cluster,
-            'fs'
-        ])
-        assert ceph_fs.generate_ceph_cmd(fake_cluster, [], image) == expected_cmd
-
     def test_create_fs(self):
         fake_module = MagicMock()
         fake_module.params = fake_params
         expected_cmd = [
             fake_binary,
+            '-n', 'client.admin',
+            '-k', '/etc/ceph/ceph.client.admin.keyring',
             '--cluster', fake_cluster,
             'fs', 'new',
             fake_fs,
@@ -90,6 +51,8 @@ class TestCephFsModule(object):
         fake_module.params = fake_params
         expected_cmd = [
             fake_binary,
+            '-n', 'client.admin',
+            '-k', '/etc/ceph/ceph.client.admin.keyring',
             '--cluster', fake_cluster,
             'fs', 'set',
             fake_fs,
@@ -104,6 +67,8 @@ class TestCephFsModule(object):
         fake_module.params = fake_params
         expected_cmd = [
             fake_binary,
+            '-n', 'client.admin',
+            '-k', '/etc/ceph/ceph.client.admin.keyring',
             '--cluster', fake_cluster,
             'fs', 'get',
             fake_fs,
@@ -117,6 +82,8 @@ class TestCephFsModule(object):
         fake_module.params = fake_params
         expected_cmd = [
             fake_binary,
+            '-n', 'client.admin',
+            '-k', '/etc/ceph/ceph.client.admin.keyring',
             '--cluster', fake_cluster,
             'fs', 'rm',
             fake_fs,
@@ -130,6 +97,8 @@ class TestCephFsModule(object):
         fake_module.params = fake_params
         expected_cmd = [
             fake_binary,
+            '-n', 'client.admin',
+            '-k', '/etc/ceph/ceph.client.admin.keyring',
             '--cluster', fake_cluster,
             'fs', 'fail',
             fake_fs
